+++ mappers/AccountMapper.ts
@@ -13,6 +13,7 @@ export class AccountMapper {
   static commercetoolsCustomerToAccount(commercetoolsCustomer: commercetoolsCustomer): Account {
     return {
       accountId: commercetoolsCustomer.id,
+      accountGroupId: commercetoolsCustomer?.customerGroup?.id,
       email: commercetoolsCustomer.email,
       salutation: commercetoolsCustomer?.salutation,
       firstName: commercetoolsCustomer?.firstName,