+++ utils/routers/ProductRouter.ts
@@ -4,7 +4,7 @@ import { ProductQuery } from '@Types/query/ProductQuery';
 import { LineItem } from '@Types/cart/LineItem';
 import { LineItem as WishlistItem } from '@Types/wishlist/LineItem';
 import { ProductApi } from '../../apis/ProductApi';
-import { getPath, getLocale, getCurrency } from '../Request';
+import { getPath, getLocale, getCurrency, getAccountGroupId } from '../Request';
 
 export class ProductRouter {
   private static isProduct(product: Product | LineItem | WishlistItem): product is Product {
@@ -35,6 +35,7 @@ export class ProductRouter {
     if (urlMatches) {
       const productQuery: ProductQuery = {
         skus: [urlMatches[1]],
+        accountGroupId: getAccountGroupId(request),
       };
       return productApi.getProduct(productQuery);
     }